+++ CHANGES.rst
@@ -1,6 +1,12 @@
 Changes
 =======
 
+Version master (UNRELEASED)
+---------------------------
+
+- Fixes `bug related to duplicated database session <https://github.com/reanahub/pytest-reana/issues/33>`_.
+- Add Black formatter support.
+
 Version 0.6.0 (2019-12-19)
 --------------------------